tests: publishers: Move zulip specific fixtures to test_zulip.py
